Klassen und Variablen umbenannt; Programmcode kommentiert.
