+++ src/main/java/at/jku/ce/adaptivetesting/questions/math/js/GeoGebraComponentState.java
@@ -1,0 +1,11 @@
+package at.jku.ce.adaptivetesting.questions.math.js;
+
+import com.vaadin.shared.ui.JavaScriptComponentState;
+
+/**
+ * Shared state for the GeoGebraComponent on the server side
+ * and the JavaScript-Component on the client side.
+ */
+public class GeoGebraComponentState extends JavaScriptComponentState {
+    public String value;
+}
