--- src/main/java/at/jku/ce/adaptivetesting/questions/math/js/MyComponentState.java
@@ -1,7 +1,0 @@
-package at.jku.ce.adaptivetesting.questions.math.js;
-
-import com.vaadin.shared.ui.JavaScriptComponentState;
-
-public class MyComponentState extends JavaScriptComponentState {
-    public String value;
-}
